Add interface class, Hitnmiss::Driver
Why you made the change:

I did this so that there would be class to aid people in making sure
they conformed to the interface of a driver if they are developing a new
driver.

How the change addresses the need:

This aids in this by raising exceptions if one of the driver methods is
used and the inheriting class neglects to implement it.