Switch from Class to Object Interface
Why you made the change:

I did this because the class-style interface made it difficult to handle
dependency injection for a repository correctly, and made testing more
difficult. This does however move the responsibility of scoping/exposing
repository instances appropriately outside of this library.

I fixed the InMemoryDriver#all method to have it correctly return the
cached values rather than the internal hash entities it uses to track
the values with expirations.