Fix some small problems in README.md
Why you made the change:

I did this so that our users would hopefully have a smoother experience
with learning about Hitnmiss and getting started with it. I also did
this to just correct statements that weren't true. I also removed the
file path comments at the top of the repository examples as I don't want
to imply/recommend a location for the cache repositories because I
believe that is a project's choice. Either the project will follow a
stance like Rails and separate things by type or a project will separate
things based on the actual application architecture. It should be
completely up to the project and Hitnmiss as a library shouldn't take
stance on it.