Reorganized a bunch of code
Why you made the change:

I extracted the get(*args) and get_all(keyspace) methods out into a
module called Fetcher and renamed them to fetch(*args) and
fetch_all(keyspace) respectively. This of course became an impetus to
rename the public interface fetch() method to get(). I did this because
I felt it provides a clearer organizational structure and aids in better
defining the interface.

I also turned the Hitnmiss::Driver into a module and moved what was the
Driver to Hitnmiss::Driver::Interface. This allowed me to then move what
were the Hitnmiss::Hit and Hitnmiss::Miss classes into the
Hitnmiss::Driver namespace which is where they should be scoped because
it is the only place that needs to know how to create them.

I also made the Hitnmiss::Fetcher module methods be private. I did this
to better communicate that they are explicitly not supposed to be part
of the public interface of the repository they are mixed into.