Modify to utilize Hit & Miss objects explicitly
Why you made the change:

I modified the InMemoryDriver and Repository to expect explicit returns
of either a Hitnmiss::Driver::Hit object or a Hitnmiss::Driver::Miss
object. This allows us to remove the ambiguous nil that we were
previously depending on. This now enables us support caching nil values
if the user wants.