Add setnoexp method to driver interface
Why you made the change:

I did this so that drivers would have to provide a mechanism to allow something
to be cached without an expiration. This will become especially important for
the concept of a interval based background refresh repository because it will
always have something cached after initial prime and will just be replacing the
cached values.