Add fingerprint to Hitnmiss::Entity
Why you made the change:

I did this so that repository fetchers would be able to pass in externally
managed fingerprint for the item that is going to be cached by Hitnmiss. This is
especially useful if you are using the HTTP ETag header on the remote source for
the item that is going be cached. Primarily, because it would then allow the
fingerprint to used later on to identify if a cache is stale or not.