Collapse driver setnoexp and set into new set
Why you made the change:

I did this because the setnoexp method seemed weird in the first place as a
secondary set, and because I was going to have to break backwards compatibility
of the set method to support fingerprents. Therefore, I decided to break the API
and make it more flexible by actually passing the cacheable entity into the set
method and giving the drivers access to the cacheable entity. This gives us more
flexibility to adjust the driver API in a backwards compatible way in the
future.