Add private `get_from_cache` method to Repository
Why you made the change:

I added this method for a couple reasons.

1. It extracts the concerns of driver lookup, key generation, and driver
interfacing out of the Repository public API `get` method.
2. It provides a nice internal method for use in implementing other Repository
public API methods. For example if we wanted to implement a public `stale?`
method on Repository, we would need to get a Hit or Miss from the driver but
wouldn't want to automatically prime the cache, etc. that is done in the public
API `get` method.