Add fingerprint support to the InMemoryDriver
Why you made the change:

I did this so that in the scenarios where a fingerprint was provided by the
fetcher Hitnmiss would make that fingerprint available internally. This enables
methods that would need this information, like a theoretically `stale?` method
to be able to do things based on fingerprint.