Switch to >= for expiration comparison
Why you made the change:

I did the because it is technically more correct as it is really trying to make
sure the appropriate time has elapsed. So, if we set an expiration of 1 second
and and time_at_which_cached + expiration = now then that time has technically
elapsed. Therefore, it should be an >= rather than just a greater >.