Remove release process from DEVELOPMENT.md
Why you made the change:

I did this because the majority of the community doesn't care about the
mechanics of our release process which is vary standard. If we maybe had some
unique things happening in our process then it might be worth having a wiki page
or writing a blog post about it. However, until then it just seems like noise.