Add logger support
Why you made the change:

I did this so that people could pass in their application logger and get
details about what is happening in Hitnmiss if they so desire on a
repository by repository basis. Given that this is using optional_logger
this also enables us to add logging to the Hitnmiss without having to
clutter our code base with a bunch of conditionals.